Add fast futex-based thread parker for Linux.
